deps/v8/test/inspector/debugger/asm-js-breakpoint-during-exec-expected.txt

This test runs asm.js which calls back to JS. JS triggers a break, on pause we set breakpoints in the asm.js code.

Running test: enableDebugger

Running test: addScript
Script nr 1 parsed!

Running test: runTestFunction
Script nr 2 parsed!
Paused #1
  - [0] {"functionName":"call_debugger","function_lineNumber":13,"function_columnNumber":24,"lineNumber":14,"columnNumber":4}
  - [1] {"functionName":"callDebugger","lineNumber":5,"columnNumber":6}
  - [2] {"functionName":"redirectFun","lineNumber":8,"columnNumber":6}
  - [3] {"functionName":"testFunction","function_lineNumber":0,"function_columnNumber":21,"lineNumber":20,"columnNumber":2}
  - [4] {"functionName":"","function_lineNumber":0,"function_columnNumber":0,"lineNumber":0,"columnNumber":0}
First time paused, setting breakpoints!
Flooding script with breakpoints for all lines (0 - 24)...
Setting breakpoint on line 0
error: undefined
Setting breakpoint on line 1
error: undefined
Setting breakpoint on line 2
error: undefined
Setting breakpoint on line 3
error: undefined
Setting breakpoint on line 4
error: undefined
Setting breakpoint on line 5
error: undefined
Setting breakpoint on line 6
error: undefined
Setting breakpoint on line 7
error: undefined
Setting breakpoint on line 8
error: undefined
Setting breakpoint on line 9
error: undefined
Setting breakpoint on line 10
error: undefined
Setting breakpoint on line 11
error: undefined
Setting breakpoint on line 12
error: undefined
Setting breakpoint on line 13
error: undefined
Setting breakpoint on line 14
error: undefined
Setting breakpoint on line 15
error: undefined
Setting breakpoint on line 16
error: undefined
Setting breakpoint on line 17
error: undefined
Setting breakpoint on line 18
error: undefined
Setting breakpoint on line 19
error: undefined
Setting breakpoint on line 20
error: undefined
Setting breakpoint on line 21
error: undefined
Setting breakpoint on line 22
error: undefined
Setting breakpoint on line 23
error: undefined
Script nr 3 parsed!
Resuming...
Paused #2
  - [0] {"functionName":"call_debugger","function_lineNumber":13,"function_columnNumber":24,"lineNumber":15,"columnNumber":2}
  - [1] {"functionName":"callDebugger","lineNumber":5,"columnNumber":6}
  - [2] {"functionName":"redirectFun","lineNumber":8,"columnNumber":6}
  - [3] {"functionName":"testFunction","function_lineNumber":0,"function_columnNumber":21,"lineNumber":20,"columnNumber":2}
  - [4] {"functionName":"","function_lineNumber":0,"function_columnNumber":0,"lineNumber":0,"columnNumber":0}
Script nr 4 parsed!
Resuming...
Paused #3
  - [0] {"functionName":"testFunction","function_lineNumber":0,"function_columnNumber":21,"lineNumber":22,"columnNumber":17}
  - [1] {"functionName":"","function_lineNumber":0,"function_columnNumber":0,"lineNumber":0,"columnNumber":0}
Script nr 5 parsed!
Resuming...
Paused #4
  - [0] {"functionName":"testFunction","function_lineNumber":0,"function_columnNumber":21,"lineNumber":23,"columnNumber":2}
  - [1] {"functionName":"","function_lineNumber":0,"function_columnNumber":0,"lineNumber":0,"columnNumber":0}
Script nr 6 parsed!
Resuming...

Running test: finished
Finished TestSuite.
